refactor: move AppField to fields file
