+++ packages/types/src/extensions/index.ts
@@ -1,4 +1,3 @@
-import type { Field } from '../fields.js';
 export * from './api-extension-context.js';
 export * from './app-extension-config.js';
 export * from './displays.js';
@@ -12,5 +11,3 @@ export * from './operations.js';
 export * from './options.js';
 export * from './panels.js';
 export * from './themes.js';
-
-export type AppField = Field & { schema: { default_value: any } };
