+++ packages/types/src/extensions/layouts.ts
@@ -5,7 +5,6 @@ export interface LayoutConfig<Options = any, Query = any> {
 	id: string;
 	name: string;
 	icon: string;
-
 	component: Component;
 	slots: {
 		options: Component;
