+++ packages/types/src/extensions/operations.ts
@@ -3,7 +3,7 @@ import type { Accountability } from '../accountability.js';
 import type { DeepPartial } from '../misc.js';
 import type { FlowRaw } from '../flows.js';
 import type { ApiExtensionContext } from './api-extension-context.js';
-import type { AppField } from './index.js';
+import type { AppField } from '../fields.js';
 
 export type OperationContext = ApiExtensionContext & {
 	data: Record<string, unknown>;
@@ -20,7 +20,6 @@ export interface OperationAppConfig {
 	name: string;
 	icon: string;
 	description?: string;
-
 	overview:
 		| ((
 				options: Record<string, any>,
@@ -37,6 +36,5 @@ export interface OperationAppConfig {
 
 export interface OperationApiConfig<Options = Record<string, unknown>> {
 	id: string;
-
 	handler: OperationHandler<Options>;
 }
